Rename relative orbital elements
