+++ s2e-ff/CMakeLists.txt
@@ -96,7 +96,7 @@ set(SOURCE_FILES
   src/Library/math/RotationFirstDualQuaternion.cpp
   src/Library/math/TranslationFirstDualQuaternion.cpp
   src/Library/Orbit/QuasiNonsingularOrbitalElements.cpp
-  src/Library/RelativeOrbit/RelativeOrbitalElements.cpp
+  src/Library/RelativeOrbit/QuasiNonsingularRelativeOrbitalElements.cpp
   src/Simulation/Case/FfCase.cpp
   src/Simulation/Spacecraft/FfSat.cpp
   src/Simulation/Spacecraft/FfComponents.cpp
